Refactor price formatting in TimerWorker to improve readability and maintainability

+++ Android/src/app/src/main/java/com/google/ai/edge/gallery/worker/TimerWorker.kt
@@ -316,9 +316,9 @@ class TimerWorker(context: Context, params: WorkerParameters) :
                     continue
                 }
 
-                val formattedPrice = BigDecimal((stockPrice * 0.99).toString())
-                    .setScale(2, RoundingMode.DOWN)
-                    .toDouble()
+                val formattedPrice =
+                    BigDecimal((stockPrice * 0.99).toString()).setScale(2, RoundingMode.DOWN)
+
 
                 // Execute Momentum Trade
                 runStep(
